增加“服务器状态感知”、项目识别与菜单优化 (#3)
## 概述
本次变更围绕「更友好地控制运行/保存」「按项目(project.json)识别」「根据服务器状态显示命令」三方面展开：
- 新增扩展上下文 autoxjs.serverStarted / autoxjs.hasProjectJson，用于动态控制菜单与操作可用性
- 支持从当前文件/路径向上查找 project.json，智能定位项目目录运行/保存
- 设备连接/断开时发出 server_started/server_stopped 事件，状态栏与上下文联动
- 优化截图功能与保存流程提示
- 修复项目打包 MD5 计算的类型问题
- 新增“启动服务器”图标资源

## 详细变更
- 新增资源
  - assets/server.svg：用于“启动服务器”命令的图标（浅/深色一致）

- package.json（配置与菜单）
  - editor/title 菜单：
    - 新增 extension.startServer（始终在 JS 文件可见，排序 group navigation@100）
- 其余命令（run/save/runProject/saveProject/captureScreen/stopAll）仅在
autoxjs.serverStarted 为真时显示，并按顺序分组（@101~@106）
- 针对 project.json
文件提供同样的菜单逻辑（startServer/runProject/saveProject/stopAll）
  - explorer/context 菜单：
    - 仅当 autoxjs.serverStarted 为真时显示
    - 文件夹：runProject（autoxjs@1）、saveProject（autoxjs@2）
    - .js 文件：run（autoxjs@3）、save（autoxjs@4）
    - 取消原 rerun 菜单，改为 run
  - commands：
    - 为 extension.startServer 增加图标 assets/server.svg
  - 注：菜单控制依赖新增上下文 autoxjs.serverStarted 与 autoxjs.hasProjectJson

- src/autojs-debug.ts（服务事件）
  - 在首个设备连接时 emit('server_started')
  - 在最后一个设备断开时 emit('server_stopped')
  - 便于 UI/菜单实时联动

- src/extension.ts（核心逻辑与 UX）
  - 状态栏
    - 初始化为“Stopped”，点击切换启动/停止
    - 基于 server_started/server_stopped 事件动态更新
  - 上下文状态
    - 初始化 setContext('autoxjs.serverStarted', false)
    - 监听 server 的 server_started/server_stopped 事件更新上下文
- 新增 autoxjs.hasProjectJson：通过当前活跃编辑器路径向上查找 project.json
判断，监听编辑器切换、保存/创建/删除/重命名事件以刷新
  - 项目识别与命令
    - 新增 findProjectFolder(url?)：从文件或目录向上查找 project.json，且不越过工作区边界
    - runProject(url?) / saveProject(url?) 优先使用找到的项目目录；找不到则回退到工作区根
    - sendProjectCommand：当未打开工作区时提示“Please open a project folder”
  - 脚本运行/保存
    - runOrRerun/saveTo：小幅清理与健壮性改进（参数、读取逻辑）
  - 截图功能
    - captureScreen：保留设备端脚本流程，VS Code 提示更新为英文；保存后自动打开图片
    - saveFileFromDevice：自动创建目录、Base64 解码保存，错误提示更清晰（英文）
  - 其他
    - 去除少量空行/空白字符，保持代码整洁

- src/project.ts（打包与 MD5）
- 修复 MD5 计算时的类型：cryto.createHash('md5').update(buffer as
Uint8Array).digest('hex')
  - 防止类型不匹配导致的校验异常

## 影响和注意事项
- 菜单显示更智能：只有在服务器已启动时才显示运行/保存相关命令；针对包含 project.json 的项目提供更准确的 run/save
项目操作
- 新增的上下文依赖事件触发：首次设备连接后才视为 serverStarted=true；全部设备断开后恢复为 false
- 截图保存默认路径为工作区根下的 screenshots/ 目录
- 英文提示替换了部分原有中文乱码提示，提升可读性

## 分类
- 配置变更：package.json 菜单与命令图标、上下文键
- 逻辑更新：服务器状态事件、项目识别、命令行为、状态栏与上下文联动、截图与文件保存
- 资源：新增 server.svg
- 其他修复：MD5 计算类型修复（project.ts）

---------

+++ src/project.ts
@@ -109,7 +109,7 @@ export class ProjectObserser {
                 });
             })
             .then(buffer => {
-                var md5 = cryto.createHash('md5').update(buffer).digest('hex');
+                var md5 = cryto.createHash('md5').update(buffer as Uint8Array).digest('hex');
                 return {
                     buffer: buffer,
                     md5: md5
